move check_if_is_orphan_chain to the service
