+++ src-tauri/src/node/node_adapter.rs
@@ -42,6 +42,13 @@ use tari_shutdown::ShutdownSignal;
 use tari_utilities::ByteArray;
 use tokio::sync::watch;
 use tokio::time::timeout;
+use serde_json::json;
+use tari_common::configuration::Network;
+use tauri_plugin_sentry::sentry;
+use tauri_plugin_sentry::sentry::protocol::Event;
+
+use crate::network_utils::{get_best_block_from_block_scan, get_block_info_from_block_scan};
+
 
 const LOG_TARGET: &str = "tari::universe::minotari_node_adapter";
 
@@ -268,6 +275,72 @@ impl NodeAdapterService {
         let mut client = BaseNodeGrpcClient::connect(self.grpc_address.clone()).await?;
         let connected_peers = client.list_connected_peers(Empty {}).await?;
         Ok(connected_peers.into_inner().connected_peers)
+    }
+
+    pub async fn check_if_is_orphan_chain(
+        &self,
+        report_to_sentry: bool,
+    ) -> Result<bool, anyhow::Error> {
+        let BaseNodeStatus {
+            is_synced,
+            block_height: local_tip,
+            ..
+        } = self.get_network_state().await?;
+        if !is_synced {
+            info!(target: LOG_TARGET, "Node is not synced, skipping orphan chain check");
+            return Ok(false);
+        }
+
+        let network = Network::get_current_or_user_setting_or_default();
+        let block_scan_tip = get_best_block_from_block_scan(network).await?;
+        let heights: Vec<u64> = vec![
+            block_scan_tip.saturating_sub(50),
+            block_scan_tip.saturating_sub(100),
+            block_scan_tip.saturating_sub(200),
+        ];
+        let mut block_scan_blocks: Vec<(u64, String)> = vec![];
+
+        for height in &heights {
+            let block_scan_block = get_block_info_from_block_scan(network, height).await?;
+            block_scan_blocks.push(block_scan_block);
+        }
+
+        let local_blocks = self.get_historical_blocks(heights).await?;
+        for block_scan_block in &block_scan_blocks {
+            if !local_blocks
+                .iter()
+                .any(|local_block| block_scan_block.1 == local_block.1)
+            {
+                if report_to_sentry {
+                    let error_msg = "Orphan chain detected".to_string();
+                    let extra = vec![
+                        (
+                            "block_scan_block_height".to_string(),
+                            json!(block_scan_block.0.to_string()),
+                        ),
+                        (
+                            "block_scan_block_hash".to_string(),
+                            json!(block_scan_block.1.clone()),
+                        ),
+                        (
+                            "block_scan_tip_height".to_string(),
+                            json!(block_scan_tip.to_string()),
+                        ),
+                        ("local_tip_height".to_string(), json!(local_tip.to_string())),
+                    ];
+                    sentry::capture_event(Event {
+                        message: Some(error_msg),
+                        level: sentry::Level::Error,
+                        culprit: Some("orphan-chain".to_string()),
+                        extra: extra.into_iter().collect(),
+                        ..Default::default()
+                    });
+                }
+                return Ok(true);
+            }
+        }
+
+        Ok(false)
     }
 }
